requirements.txt
pandas
requests
python-dotenv
tqdm
scikit-learn
joblib
matplotlib
seaborn
notebook
xgboost